Create a file for find_default_config_files
