+++ .isort.cfg
@@ -1,7 +1,7 @@
 [settings]
 multi_line_output=3
 line_length=88
-known_third_party=astroid, sphinx, isort, pytest, mccabe, six,
+known_third_party=astroid, sphinx, isort, pytest, mccabe, six, toml
 include_trailing_comma=True
 skip_glob=tests/functional/**,tests/input/**,tests/extensions/data/**,tests/regrtest_data/**,tests/data/**,astroid/**,venv/**
 project=pylint
